detect/http_raw_host: move raw into regular host logic
